Split mod.rs itself at clean top-level boundaries to clear max-lines
